+++ packages/riverpod/lib/src/framework/family.dart
@@ -33,11 +33,11 @@ typedef FamilyCreate<T, R extends Ref, Arg> = T Function(R ref, Arg arg);
 /// A base class for all families
 abstract class Family<
     @Deprecated(
-  'The generic parameter will be removed in version 3.0.0. '
-  'This is to enable riverpod_generator to implement families with generic parameters',
-)
-        // ignore: deprecated_member_use_from_same_package
-        State> implements FamilyOverride<State>, ProviderOrFamily {
+      'The generic parameter will be removed in version 3.0.0. '
+      'This is to enable riverpod_generator to implement families with generic parameters',
+    )
+    // ignore: deprecated_member_use_from_same_package
+    State> implements FamilyOverride<State>, ProviderOrFamily {
   /// A base class for all families
   const Family();