+++ packages/riverpod/lib/src/future_provider/base.dart
@@ -88,7 +88,7 @@ class FutureProvider<T> extends _FutureProviderBase<T>
 class FutureProviderElement<T> extends ProviderElementBase<AsyncValue<T>>
     with FutureHandlerProviderElementMixin<T>
     implements FutureProviderRef<T> {
-  FutureProviderElement._(_FutureProviderBase<T> super.provider);
+  FutureProviderElement._(_FutureProviderBase<T> super._provider);
 
   @override
   Future<T> get future {
@@ -112,7 +112,7 @@ class FutureProviderFamily<R, Arg> extends FamilyBase<FutureProviderRef<R>,
     AsyncValue<R>, Arg, FutureOr<R>, FutureProvider<R>> {
   /// The [Family] of a [FutureProvider]
   FutureProviderFamily(
-    super.create, {
+    super._createFn, {
     super.name,
     super.dependencies,
   }) : super(
@@ -125,7 +125,7 @@ class FutureProviderFamily<R, Arg> extends FamilyBase<FutureProviderRef<R>,
   /// Implementation detail of the code-generator.
   @internal
   FutureProviderFamily.generator(
-    super.create, {
+    super._createFn, {
     required super.name,
     required super.dependencies,
     required super.allTransitiveDependencies,