+++ packages/riverpod/test/framework/auto_dispose_test.dart
@@ -318,7 +318,7 @@ final alwaysAlive = Provider((ref) {
     final onDispose = OnDisposeMock();
     final dep = StateProvider((ref) => 0);
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       return ref.watch(dep);
     });
 
@@ -407,7 +407,8 @@ final alwaysAlive = Provider((ref) {
     final root = createContainer();
     final container = createContainer(parent: root);
 
-    final sub = container.listen(provider, listener, fireImmediately: true);
+    final sub =
+        container.listen(provider, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(null, 0));
     expect(buildCount, 1);
@@ -423,7 +424,7 @@ final alwaysAlive = Provider((ref) {
     expect(container.getAllProviderElements(), isEmpty);
 
     value = 42;
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     expect(buildCount, 2);
     expect(disposeCount, 1);
@@ -529,6 +530,26 @@ final alwaysAlive = Provider((ref) {
     expect(container.getAllProviderElements(), isEmpty);
   });
 
+  test('supports disposing of overridden families', () async {
+    // Regression test for https://github.com/rrousselGit/riverpod/issues/2480
+    final provider = Provider.autoDispose.family<int, int>((ref, _) => -1);
+
+    var constructionCount = 0;
+    final root = createContainer();
+    final container = createContainer(
+      parent: root,
+      overrides: [provider.overrideWith((ref, arg) => ++constructionCount)],
+    );
+
+    var count = container.read(provider(0));
+    expect(count, 1);
+
+    await container.pump();
+
+    count = container.read(provider(0));
+    expect(count, 2);
+  });
+
   test(
       'can select auto-dispose providers if the selecting provider is auto-dispose too',
       () {
@@ -547,7 +568,7 @@ final alwaysAlive = Provider((ref) {
     late AutoDisposeRef<Object?> ref;
     final provider = Provider.autoDispose((r) {
       ref = r;
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       ref.maintainState = true;
     });
     final container = createContainer();
@@ -575,14 +596,15 @@ final alwaysAlive = Provider((ref) {
     var value = 42;
     final onDispose = OnDisposeMock();
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       ref.maintainState = true;
       return value;
     });
     final container = createContainer();
     final listener = Listener<int>();
 
-    final sub = container.listen(provider, listener, fireImmediately: true);
+    final sub =
+        container.listen(provider, listener.call, fireImmediately: true);
     verify(listener(null, 42)).called(1);
     verifyNoMoreInteractions(listener);
     sub.close();
@@ -592,7 +614,7 @@ final alwaysAlive = Provider((ref) {
     verifyZeroInteractions(onDispose);
 
     value = 21;
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     verify(listener(null, 42)).called(1);
     verifyNoMoreInteractions(listener);
@@ -616,12 +638,12 @@ final alwaysAlive = Provider((ref) {
     final container = createContainer();
     final aDispose = OnDisposeMock();
     final a = Provider.autoDispose((ref) {
-      ref.onDispose(aDispose);
+      ref.onDispose(aDispose.call);
       return 42;
     });
     final bDispose = OnDisposeMock();
     final b = Provider.autoDispose((ref) {
-      ref.onDispose(bDispose);
+      ref.onDispose(bDispose.call);
       ref.watch(a);
       return '42';
     });
@@ -650,17 +672,17 @@ final alwaysAlive = Provider((ref) {
     final onDispose = OnDisposeMock();
     var value = 42;
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       return value;
     });
     final onDispose2 = OnDisposeMock();
     final provider2 = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose2);
+      ref.onDispose(onDispose2.call);
       return ref.watch(provider);
     });
     final listener = Listener<int>();
 
-    var sub = container.listen(provider2, listener, fireImmediately: true);
+    var sub = container.listen(provider2, listener.call, fireImmediately: true);
 
     verify(listener(null, 42)).called(1);
     verifyNoMoreInteractions(listener);
@@ -684,7 +706,7 @@ final alwaysAlive = Provider((ref) {
     verifyNoMoreInteractions(onDispose2);
 
     value = 21;
-    sub = container.listen(provider2, listener, fireImmediately: true);
+    sub = container.listen(provider2, listener.call, fireImmediately: true);
 
     verify(listener(null, 21)).called(1);
     verifyNoMoreInteractions(listener);
@@ -696,12 +718,12 @@ final alwaysAlive = Provider((ref) {
     final container = createContainer();
     final aDispose = OnDisposeMock();
     final a = Provider.autoDispose((ref) {
-      ref.onDispose(aDispose);
+      ref.onDispose(aDispose.call);
       return 42;
     });
     final bDispose = OnDisposeMock();
     final b = Provider.autoDispose((ref) {
-      ref.onDispose(bDispose);
+      ref.onDispose(bDispose.call);
       return 42;
     });
 
@@ -733,7 +755,7 @@ final alwaysAlive = Provider((ref) {
     final container = createContainer();
     final onDispose = OnDisposeMock();
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       return 42;
     });
 
@@ -758,7 +780,7 @@ final alwaysAlive = Provider((ref) {
     final container = createContainer();
     final onDispose = OnDisposeMock();
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       return 42;
     });
 
@@ -787,7 +809,7 @@ final alwaysAlive = Provider((ref) {
     final container = createContainer();
     final onDispose = OnDisposeMock();
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       return 42;
     });
 
@@ -812,7 +834,7 @@ final alwaysAlive = Provider((ref) {
       () async {
     final onDispose = OnDisposeMock();
     final provider = Provider.autoDispose((ref) {
-      ref.onDispose(onDispose);
+      ref.onDispose(onDispose.call);
       return 42;
     });
     final container = createContainer();