+++ packages/riverpod/test/providers/stream_provider/stream_provider_test.dart
@@ -186,7 +186,7 @@ void main() {
       return Stream.value(0);
     });
 
-    container.listen(provider, listener);
+    container.listen(provider, listener.call);
     await container.read(provider.future);
 
     expect(ref.state, const AsyncData<int>(0));
@@ -257,7 +257,7 @@ void main() {
     );
 
     container.read(dep.notifier).state = controller.stream;
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     verifyOnly(
       listener,
@@ -642,7 +642,7 @@ void main() {
     });
     final listener = Listener<AsyncValue<int>>();
 
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(null, const AsyncValue.loading()));
 
@@ -663,7 +663,7 @@ void main() {
     final listener = Listener<Stream<int>>();
 
     // ignore: deprecated_member_use_from_same_package
-    container.listen(provider.stream, listener, fireImmediately: true);
+    container.listen(provider.stream, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(any, any));
 
@@ -684,7 +684,7 @@ void main() {
     });
     final listener = Listener<Future<int>>();
 
-    container.listen(provider.future, listener, fireImmediately: true);
+    container.listen(provider.future, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(any, any));
 
@@ -820,7 +820,7 @@ void main() {
     final provider = StreamProvider((_) => controller.stream);
     final listener = Listener<AsyncValue<int>>();
 
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(null, const AsyncValue<int>.loading()));
 
@@ -849,7 +849,7 @@ void main() {
     final error = Error();
     final stack = StackTrace.current;
 
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(null, const AsyncValue<int>.loading()));
 
@@ -883,12 +883,12 @@ void main() {
     final controller = StreamController<int>(sync: true);
     final dispose = OnDisposeMock();
     final provider = StreamProvider((ref) {
-      ref.onDispose(dispose);
+      ref.onDispose(dispose.call);
       return controller.stream;
     });
     final listener = Listener<AsyncValue<int>>();
 
-    container.listen(provider, listener, fireImmediately: true);
+    container.listen(provider, listener.call, fireImmediately: true);
 
     verifyOnly(listener, listener(null, const AsyncValue<int>.loading()));
 
@@ -947,7 +947,7 @@ void main() {
       final sub = container.listen(
         // ignore: deprecated_member_use_from_same_package
         streamProvider.stream,
-        listener,
+        listener.call,
         fireImmediately: true,
       );