+++ packages/riverpod_lint/lib/src/assists/stateless_to_stateful_provider.dart
@@ -2,7 +2,7 @@ import 'package:analyzer/source/source_range.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../riverpod_custom_lint.dart';
-import 'convert_to_consumer_widget.dart';
+import 'stateful_to_stateless_provider.dart';
 
 class StatelessToStatefulProvider extends RiverpodAssist {
   StatelessToStatefulProvider();