+++ website/i18n/it/docusaurus-plugin-content-docs/current/cookbooks/testing_original_test_dart.dart
@@ -27,7 +27,7 @@ void main() {
     // Osserva un provider e ne spia i cambiamenti
     container.listen<int>(
       counterProvider,
-      listener,
+      listener.call,
       fireImmediately: true,
     );
 
@@ -52,7 +52,7 @@ void main() {
 
     container.listen<int>(
       counterProvider,
-      listener,
+      listener.call,
       fireImmediately: true,
     );